+++ src/main/kotlin/org/mybatis/dynamic/sql/util/kotlin/JoinCollector.kt
@@ -16,8 +16,8 @@
 package org.mybatis.dynamic.sql.util.kotlin
 
 import org.mybatis.dynamic.sql.BindableColumn
+import org.mybatis.dynamic.sql.RenderableCondition
 import org.mybatis.dynamic.sql.SqlBuilder
-import org.mybatis.dynamic.sql.VisitableCondition
 
 typealias JoinReceiver = JoinCollector.() -> Unit
 
@@ -38,7 +38,7 @@ class JoinCollector {
     }
 }
 
-class RightColumnCollector<T : Any>(private val joinConditionConsumer: (VisitableCondition<T>) -> Unit) {
+class RightColumnCollector<T : Any>(private val joinConditionConsumer: (RenderableCondition<T>) -> Unit) {
     infix fun equalTo(rightColumn: BindableColumn<T>) = joinConditionConsumer.invoke(SqlBuilder.isEqualTo(rightColumn))
 
     infix fun equalTo(value: T) = joinConditionConsumer.invoke(SqlBuilder.isEqualTo(value))